Argus/requirements.txt
flask==3.0.0

python-nmap==0.7.1
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.3.0
nvdlib==0.5.4
colorama==0.4.6
python-dotenv==1.0.0
joblib==1.3.2
scikit-learn==1.3.0
